Lock to docker 1.7.0, fix ecs tcs mounts
After docker 1.7.0 we (Remind) have seen this issue pop up:

https://github.com/docker/docker/issues/13914

Now that this is closed:

https://github.com/docker/docker/pull/16001

We can use the new docker-engine packages, but with an old version. I'm
moving the empire_ami back to docker 1.7.0 till 13914 above is fixed.

As well, this should fix ECS stats - we just needed a bunch of volumes,
per:

https://github.com/aws/amazon-ecs-agent/issues/174

Doing this internally @ Remind fixed this.