Add a fork of docker.conf with a patch to handle docker.sock being a directory
There are cases where /var/run/docker.sock can get mounted as a
directory due to a race condition when restarting the docker daemon and
docker containers that mount /var/run/docker.sock. If the container
attempts to mount /var/run/docker.sock and it doesn't exist, it will
default it to a directory (similar to this thread:
https://bugzilla.redhat.com/show_bug.cgi?id=1289851). We can't use
`--restart=always` since it doesn't allow you to provide a name for the
container, which means we wouldn't be able to use upstart to stop the
container by name.